clear compiler error

+++ silk/src/pin.hpp
@@ -5,6 +5,6 @@
 struct Pin {
   float pin_sitffness = 1e6f;
 
-  Eigen::VectorXf index;
+  Eigen::VectorXi index;
   Eigen::VectorXf position;
 };
